[refactoring] some header and sample cleanup
Fixes some headers not working but themselves.

Cleans up examples and to show a different way of using them.
