+++ samples/demo/demo.h
@@ -5,7 +5,13 @@
 #include <string>
 #include <memory>
 
+#ifdef FCTLOG_ENABLE
+#include "fctlog/define.h"
+#endif
+
 namespace fctlog {
+
+namespace internal {
 
 class Demo {
 public:
@@ -18,5 +24,28 @@ public:
   virtual int f2Const(int x, int y) const;
   virtual int f3(std::unique_ptr<int>& foo) const;
 };
+
+#ifdef FCTLOG_ENABLE
+#define FCTLOG_PARENT_CLASS Demo
+class DemoLog : public Demo {
+public:
+  FCTLOG_METHOD0(void, fv);
+  FCTLOG_METHOD0(int, f0);
+  FCTLOG_CONST_METHOD0(int, f0Const);
+  FCTLOG_METHOD1(int, f1, (int val = 0), val);
+  FCTLOG_METHOD2(int, f2, (int val = 0, const std::string &str = ""), val, str);
+  FCTLOG_CONST_METHOD2(int, f2Const, (int x, int y), x, y);
+  FCTLOG_CONST_METHOD1(int, f3, (std::unique_ptr<int>& foo), foo);
+};
+#undef FCTLOG_PARENT_CLASS
+#endif
+
+} // namespace internal
+
+#ifdef FCTLOG_ENABLE
+using Demo = internal::DemoLog;
+#else
+using Demo = internal::Demo;
+#endif
 
 } // namespace fctlog
