+++ samples/template/demo.h
@@ -5,7 +5,11 @@
 #include <string>
 #include <vector>
 
-namespace fctlog {
+#ifdef FCTLOG_ENABLE
+#include "fctlog/define.h"
+#endif
+
+namespace fctlog { namespace internal {
 
 template <typename T1, typename T2>
 class Demo {
@@ -15,5 +19,26 @@ public:
   virtual int f1(T1 x) { return x; }
   virtual T1 f2Const(T2 x, T2 y) const { return x + y; }
 };
+
+#ifdef FCTLOG_ENABLE
+#define FCTLOG_PARENT_CLASS Demo<T1, T2>
+template <typename T1, typename T2>
+class DemoLog : public Demo<T1, T2> {
+public:
+  FCTLOG_METHOD1(void, fv, (const std::vector<T1> &v), v);
+  FCTLOG_METHOD1(int, f1, (T1 x), x);
+  FCTLOG_CONST_METHOD2(T1, f2Const, (T2 x, T2 y), x, y);
+};
+#undef FCTLOG_PARENT_CLASS
+#endif
+}
+
+#ifdef FCTLOG_ENABLE
+template<typename T1, typename T2>
+using Demo = internal::DemoLog<T1, T2>;
+#else
+template<typename T1, typename T2>
+using Demo = internal::Demo<T1, T2>;
+#endif
 
 } // namespace fctlog
